alteraçoes.

+++ sistemadepecasbg/src/br/edu/facear/sistemadepecasbg/dao/Dao.java
@@ -8,6 +8,6 @@ public interface Dao<T> {
 	public void editar(T t);	
 	public void excluir(T t);	
 	public List<T> listar();
-	public T buscar(Long id);
+	public T buscar(Integer id);
 
 }
